Exercise 462 and 463
Original version of Exercise 462 is trival, since we assume length of
equation is the same as length of solution.

However, in 463 the length of solution may be bigger than solution.

Therefore, we must change the template and implementation of plug-in
func.

And we have to input reverse of solution and equation to plug-in func.

Tested-by: ihaaae <5929624d@gmail.com>